Note deletion animation + fix layout issues (#67)
* add gap between btn. and title

* change deletion title layout + show confirmation message before deletion

* create pop scope widget + refactor

* deletion animation

* fix text overflow

* replace _buildTile with _NoteListTileWrapper widget

* update current collection after edit

* center page title in selection state + reject back btn. event on del state

+++ lib/views/home/bloc.dart
@@ -101,8 +101,15 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
         }
         final currentIdleState = switch (state) {
           IdleState state => state,
-          NonIdleState(:final previousState) => previousState,
+          NonIdleState(
+            :final previousState,
+          ) =>
+            previousState,
+          _ => null,
         };
+        if (currentIdleState case null) {
+          return;
+        }
         final showNotes = switch (event.index) {
           0 || 1 => event.index == 0,
           _ => null,
@@ -242,30 +249,39 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
     );
     on<DeleteNotesEvent>(
       (event, emit) async {
+        if (state case DeleteItemsState()) {
+          return;
+        }
         switch (state) {
           case final SelectNotesState state:
             final future = repository.removeNotes(
               state.itemIds,
             );
             emit(
-              DeleteItemsState(
-                previousState: state.previousState,
+              DeleteNotesState(
+                previousState: state,
                 future: future,
               ),
             );
+            await Future.delayed(
+              const Duration(
+                milliseconds: 500,
+              ),
+            );
+            final notes = state.previousState.notes.indexed.where(
+              (element) {
+                final (i, _) = element;
+                return !state.selected[i];
+              },
+            ).map(
+              (element) {
+                final (_, e) = element;
+                return e;
+              },
+            ).toList();
             emit(
               state.previousState.copyWith(
-                notes: state.previousState.notes.indexed.where(
-                  (element) {
-                    final (i, _) = element;
-                    return !state.selected[i];
-                  },
-                ).map(
-                  (element) {
-                    final (_, e) = element;
-                    return e;
-                  },
-                ).toList(),
+                notes: notes,
               ),
             );
             await future;
